Task #264: Order same-day sales chronologically (latest first) on Sales History
Operator complaint: same-date sales returned in random order on every page
reload, making it hard to follow lot chronology / read the new "Remaining #
Bags" column down a single date.

Root cause: getSalesHistory only ordered by soldAt, and every write path
anchors soldAt to noon IST of the picked date — so same-date sales had
identical soldAt instants and Postgres was free to reorder them.

Operator iterated on the approach during planning; final agreed design uses
soldAt + a new createdAt column + Receipt # (lot_no) numeric:
- shared/schema.ts: add `createdAt timestamptz NOT NULL DEFAULT NOW()` to the
  salesHistory table; omit it from insertSalesHistorySchema so it's strictly
  server-managed (matches the maintenanceRecords pattern). Heavy comment
  documents the IST timezone contract and the AT TIME ZONE wrapping
  requirement for any future date-portion read.
- server/migrations.ts: new entry "2026-05-01_add_created_at_to_sales_history"
  — ADD COLUMN, backfill created_at = sold_at for legacy rows, then SET
  DEFAULT NOW() and SET NOT NULL. Idempotent. Hostinger VPS picks it up on
  first boot via the existing in-house migration runner. Verified locally:
  45/45 rows backfilled, none NULL.
- server/storage.ts:1849 (getSalesHistory): orderBy chain is
  `desc(soldAt), desc(createdAt), NULLIF(regexp_replace(lot_no,'[^0-9]','','g'))::bigint ASC NULLS LAST`.
  Cast to BIGINT (not INTEGER) — caught during validation: the
  null-cs-bill-coexist smoke script seeds fixtures with timestamp-based
  18-digit lot numbers that overflow int4. bigint (int8) handles them and
  is still trivially correct for real receipt #s in the small-int range.
- client/src/pages/SalesHistory.tsx:874 (facilitatedSales comparator):
  rewritten to mirror the server's three-key chain. Uses .getTime() on
  soldAt/createdAt (TZ-neutral) and parseInt+lexicographic fallback on lotNo.

IST correctness: leverages the existing connection-level
SET TIME ZONE 'Asia/Kolkata' in server/db.ts, so DEFAULT NOW() captures IST
wall-clock automatically. No app-side new Date() is used to set createdAt,
avoiding clock-skew poisoning.

Verified after the int4→int8 fix:
- Migration applied cleanly on first restart.
- Synthetic test: bumping one row's created_at to NOW() correctly hoisted it
  to top of its date group; restored cleanly.
- Numeric receipt# tiebreaker works (lot_no 15 sorts before 100, not after).
- App preview loads without errors.
- Both safety scripts pass: null-cs-bill-coexist (7/7), sale-invalidation.

Plan file: the user revised this task several times during planning. The
current authoritative plan is in BOTH .local/tasks/task-264.md and
.local/tasks/sales-history-stable-sort-by-id.md (kept in sync).

Follow-up #265 proposed: regression script for the new sort order, mirroring
the existing safety-script pattern.

+++ server/storage.ts
@@ -1849,8 +1849,13 @@ export class DatabaseStorage implements IStorage {
     //                             For legacy rows the migration backfilled
     //                             createdAt = soldAt, so legacy collisions
     //                             still tie here and fall through to (3).
-    //   3. lot_no ASC, numeric  — receipt # tiebreaker, cast to integer so
+    //   3. lot_no ASC, numeric  — receipt # tiebreaker, cast to bigint so
     //                             "2" sorts before "11" before "100".
+    //                             bigint (int8) is used instead of int4
+    //                             because some smoke-test/fixture rows use
+    //                             timestamp-based lot numbers that overflow
+    //                             a 32-bit signed integer; real receipt #s
+    //                             are small but we keep room defensively.
     //                             regexp_replace strips any non-numeric
     //                             chars defensively; NULLIF keeps an empty
     //                             string from blowing up the cast and the
@@ -1864,7 +1869,7 @@ export class DatabaseStorage implements IStorage {
       .orderBy(
         desc(salesHistory.soldAt),
         desc(salesHistory.createdAt),
-        sql`NULLIF(regexp_replace(${salesHistory.lotNo}, '[^0-9]', '', 'g'), '')::integer ASC NULLS LAST`,
+        sql`NULLIF(regexp_replace(${salesHistory.lotNo}, '[^0-9]', '', 'g'), '')::bigint ASC NULLS LAST`,
       );
 
     if (sales.length === 0) return [];
